Wrote a basic unit test for the tractconverter
The only unit test checks if the streamlines are the same after
converting from one format to another one. Moreover, it checks if the
the conversion from the converted file back to the original is working.

NB: checks are only made on the data points, not the header nor the
attributes.